Stronger inference in "if" test.
If every block of statements in an "if" returns, then the rest of the code is
effectively an "else", so the same type inference that was prepared for the
"else" block can be used throughout.

Example:

```
local function f() : integer|string
    return "foo"
end

local x = f()

if type(x) == "integer" then
   return x
end

print((x):format())
```

If we comment `return x` above, we get a type error.

Note that for this inference to be valid, every if/elseif/else block must end
with a return, and there must be no goto's inside them.

As a special case, the standard global function error() is also detected as a
form of return.

The motivation was to be able do this:

```
local obj, err = get_object()
if not obj then error("error getting object: "..err, nil) end

-- from here on I can assume that obj is my object! :-)
```