Only apply filter if type is found.
I think this is a bugfix for #52. Please verify.

The only change in logic is adding if's after get_index.

The rest of the diff is there just because of changed
variable names.

I had to add a local `var` to make sure that `l[name]` is
updated only if get_index succeeds.

I also changed the variable names from i to idx because
luacheck noted that it was shadowing the i variable from
the outer for loop. I thought this would make the code
easier to read.