Improve inference of "and".
The goal here was to make the "ternary operator" idiom `x = y and z or w` work.

Test case:
```lua
local y = 1
local x = y > 9 and "yes" or "no"
print(string.match(x, "yes"))
```

With this patch, x is correctly inferred to be `string`.

The idea is that, in `A and B`, if A is inferred to be a boolean,
then the only boolean it can be is false, because if it is true,
then the expression will return B.

With the simplicity of this fix, I begin to fully appreciate
the elegance of having literal types in the language!!