Safer env for recursives + variable identity.
The change with the "@" characters is a bit of paranoia, as different
incompatible subtypes may concatenate to the same string ("aa".."b" and
"a".."ab") and this may cause evaluation errors in the future.

Note the change in `subtype_variable` for when the two entries have equal
names. This is the important one. Please verify if this is semantically
correct. Here it seems to solve #55, but I don't know if that's enough.